FIX: type conversion warning
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@795 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Base/cmtkDataGridMorphologicalOperators.cxx
@@ -299,7 +299,7 @@ DataGrid::GetBoundaryMap( const bool multiValued ) const
 	if ( bp )
 	  {
 	  if ( multiValued )
-	    boundary[offset] = value;
+	    boundary[offset] = static_cast<unsigned short>( value );
 	  else
 	    boundary[offset] = 1;
 	  }
